+++ Tests/Runtime/WrappedList.cs
@@ -45,22 +45,5 @@ namespace UnderLogic.Serialization.Toml.Tests
         public static WrappedList<T> Null() => new();
 
         public static WrappedList<T> FromValues(params T[] values) => new(values);
-
-        public bool AreElementsSame(IReadOnlyList<T> other)
-        {
-            if (other == null)
-                return false;
-
-            if (Count != other.Count)
-                return false;
-
-            for (var i = 0; i < Count; i++)
-            {
-                if (!this[i].Equals(other[i]))
-                    return false;
-            }
-
-            return true;
-        }
     }
 }