+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2022-12-27
+
+### Added
+
+- Unit tests for bidirectional serialization to and from TOML
+
+### Fixed
+
+- Serialization of arrays and lists containing `null` values
+
 ## [0.3.1] - 2022-12-26
 
 ### Fixed